fix(indexer): decouple onchain seed account loading
