+++ packages/indexer/src/reconcile.ts
@@ -18,6 +18,13 @@ import {
   ReconciliationCheck,
 } from "./internal/reconciliation";
 import { parseIndexerPowerSource } from "./handler/token";
+import {
+  loadKnownTokenAccounts,
+  QueryableDataSource,
+} from "./onchain-refresh/known-accounts";
+
+export { loadKnownTokenAccounts };
+export type { QueryableDataSource };
 
 interface ReconcileCliOptions {
   configPath: string;
@@ -68,13 +75,6 @@ interface ProposalReconciliationResult {
   voteSamples: VotePowerSampleResult[];
 }
 
-export interface QueryableDataSource {
-  query(sql: string, parameters?: unknown[]): Promise<any[]>;
-  transaction?<T>(
-    callback: (entityManager: QueryableDataSource) => Promise<T>
-  ): Promise<T>;
-}
-
 export interface OnchainPowerReconcileOptions {
   chainId: number;
   daoCode?: string | null;
@@ -93,8 +93,6 @@ interface OnchainPowerReconcileResult {
   balancesUpdated: number;
   powersUpdated: number;
 }
-
-const zeroAddress = "0x0000000000000000000000000000000000000000";
 
 function isHistoricalVoteUnavailable(error: unknown): boolean {
   const message =
@@ -134,82 +132,6 @@ function deriveReconcilePowerTimepoint(
   }
 
   return options.blockNumber ?? blockNumber;
-}
-
-export async function loadKnownTokenAccounts(
-  dataSource: QueryableDataSource,
-  options: OnchainPowerReconcileOptions
-): Promise<string[]> {
-  const rows = await dataSource.query(
-    `
-      WITH known_accounts AS (
-        SELECT id AS account
-        FROM contributor
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT "from" AS account
-        FROM delegate_mapping
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT "to" AS account
-        FROM delegate_mapping
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT "from" AS account
-        FROM token_transfer
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT "to" AS account
-        FROM token_transfer
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT voter AS account
-        FROM vote_cast
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT voter AS account
-        FROM vote_cast_group
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT delegator AS account
-        FROM delegate_changed
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT from_delegate AS account
-        FROM delegate_changed
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT to_delegate AS account
-        FROM delegate_changed
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-        UNION
-        SELECT delegate AS account
-        FROM delegate_votes_changed
-        WHERE chain_id = $1
-          AND lower(governor_address) = lower($2)
-      )
-      SELECT DISTINCT lower(account) AS account
-      FROM known_accounts
-      WHERE account IS NOT NULL
-        AND lower(account) <> $3
-      ORDER BY account ASC
-    `,
-    [options.chainId, options.governorAddress, zeroAddress]
-  );
-
-  return rows
-    .map((row) => normalizeAddress(row.account))
-    .filter((account): account is string => Boolean(account));
 }
 
 async function readContributorSnapshot(
@@ -657,10 +579,6 @@ function toBigInt(value: string | number | bigint | null | undefined): bigint {
   }
 
   return BigInt(value);
-}
-
-function normalizeAddress(value: string | null | undefined): string | undefined {
-  return value ? value.toLowerCase() : undefined;
 }
 
 async function createDatabaseConnection(): Promise<DataSource> {
